move resizeImage inside upload-blocks
